админка без контакт и нет комментов

+++ admin/controllers/UsersadmController.php
@@ -7,13 +7,32 @@ class UsersadmController implements Controller
     use errTrait;
     use ResponseTrait;
     use UploadTrait;
+    use ExtraTrait;
+
+    const TABLE_NAME = 'users';
+    const PAGE_NAME = 'users-adm/index';
 
     public function index()
     {
         global $smarty;
+        $pageName = self::PAGE_NAME; //для пагинации
+        $siteData = InfoModel::info();//info сайта
+        $userLimit = $siteData['pageLimitUserPanel'];// колво статей на странице
+        $tableName = self::TABLE_NAME;
+        $data = $this->pagination($pageName, $userLimit, $tableName);
+
         $user = new UsersadmModel();
+        $user->page = $data['page'];
+        $user->start = $userLimit * ($data['page'] - 1);//LIMIT start
+        $user->limit = $userLimit;//LIMIT finish
         $users = $user->all();
+        $hint = $this->hint($users, 'username');
+        $tagSearch = $this->hint($users, 'login');
+        $smarty->assign('tagSearch', $tagSearch);
+        $smarty->assign('hint', $hint);
         $smarty->assign('users', $users);
+        $smarty->assign('userLimit', $userLimit);
+        $smarty->assign('pagination', $data['pagination']);//пагинация
         $smarty->display('admin/users.tpl');
     }
 
@@ -35,10 +54,14 @@ class UsersadmController implements Controller
         if (!empty($_FILES)) {
             $uploads_dir = '../public/pic/avatar'; //дериктория куда сохраняем
             $oldIcoDir = $uploads_dir . '/' . $userData[$column];
-            if (file_exists($oldIcoDir)) {
+            $err = $this->getErrImg();
+            if ($err) {
+                $this->getResponse(['success' => false, 'err' => $err, 'name_pic_old' => $userData[$column]]);
+            }
+            $name_pic = $this->uploadPic($uploads_dir);//трейт загрузки
+            if (file_exists($oldIcoDir) && $oldIcoDir !== $uploads_dir . '/' . 'anonimus.png') {
                 unlink($oldIcoDir);//удаляем старый фаил
             };
-            $name_pic=$this->uploadPic($uploads_dir);//трейт загрузки
             $users = new UsersadmModel();
             $users->text = $name_pic;
             $users->id = $id;
@@ -50,19 +73,19 @@ class UsersadmController implements Controller
             $text = filter_var(trim($_POST['text']), FILTER_SANITIZE_STRING);
             switch (true) {
                 case $column === 'username':
-                    $err = $this->getErrUser($login = null, $pass = null, $rePass = null, $text, $mail = null, $imgFile = null);
+                    $err = $this->getErrUser($login = null, $pass = null, $rePass = null, $text, $mail = null);
                     if ($err) {
                         $this->getResponse(['success' => false, 'err' => $err]);
                     }
                     break;
                 case $column === 'email':
-                    $err = $this->getErrUser($login = null, $pass = null, $rePass = null, $userName = null, $text, $imgFile = null);
+                    $err = $this->getErrUser($login = null, $pass = null, $rePass = null, $userName = null, $text);
                     if ($err) {
                         $this->getResponse(['success' => false, 'err' => $err]);
                     }
                     break;
                 case $column === 'login':
-                    $err = $this->getErrUser($text, $pass = null, $rePass = null, $userName = null, $mail = null, $imgFile = null);
+                    $err = $this->getErrUser($text, $pass = null, $rePass = null, $userName = null, $mail = null);
                     if ($err) {
                         $this->getResponse(['success' => false, 'err' => $err]);
                     }
@@ -80,7 +103,7 @@ class UsersadmController implements Controller
 
     public function removed()
     {
-        $id = $_POST['id'];
+        $id = filter_var(trim($_POST['id']), FILTER_SANITIZE_STRING);
         if (!empty($id)) {
             $checkUser = new UserModel();
             $checkUser = $checkUser->getUserByLogin($login = null, $id);
@@ -91,8 +114,17 @@ class UsersadmController implements Controller
             $removedUser = new UsersadmModel();
             $removedUser->id = $id;
             $removedUser = $removedUser->removed();
-            $this->getResponse(['success' => $removedUser, 'err' => 'Пользователь не был удален из БД, пожалуйста очистьте 
+            if (!$removedUser) {
+                $this->getResponse(['success' => $removedUser, 'err' => 'Пользователь не был удален из БД, пожалуйста очистьте 
             кеш, обновите страницу. Если ошибка повторится свяжитесть с администратором сайта.']);
+            }
+
+            $uploads_dir = '../public/pic/avatar'; //дериктория куда сохраняем
+            $avatar = $uploads_dir . '/' . $checkUser['avatar'];
+            if (file_exists($avatar) && $avatar !== $uploads_dir . '/' . 'anonimus.png') {
+                unlink($avatar);//удаляем старый фаил
+            };
+            $this->getResponse(['success' => $removedUser]);
         }
         $this->getResponse(['success' => false, 'err' => 'Пришел пустой запрос на удаление, пожалуйста очистьте кеш, 
         обновите страницу. Если ошибка повторится свяжитесть с администратором сайта.']);
